\name{hsicCCA-package}
\alias{hsicCCA-package}
\docType{package}
\title{
Canonical Correlation Analysis based on Kernel Independence Measures
}
\description{
Canonical correlation analysis that extracts nonlinear correlation through the use of Hilbert Schmidt Independence Criterion and Centered Kernel Target Alignment.
}
\details{
\tabular{ll}{
Package: \tab hsicCCA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-13\cr
License: \tab GPL-2\cr
}

}
\author{

Billy Chang: <billy.chang@mail.utoronto.ca>

}
\references{
Chang et. al. (2013) Canonical Correlation Analysis based on Hilbert-Schmidt Independence Criterion and Centered Kernel Target Alignment. ICML 2013.

Gretton et. al. (2005) Measuring statistical dependence with Hilbert-Schmidt Norm. In Algorithmic Learning Theory 2005.

Cortes et. al. (2012) Algorithms for learning kernels based on centered alignments. JMLR 13:795-828.
}

\keyword{ package }
